Outsource a test of TitleHeader in Profile page.
